Fix error when song list is empty
If there are no songs in the regular song list, and the user mounts up
on a mount that has no custom songs defined for it, an error will be
thrown as we try to take a random song from an empty list.

To fix it we just add a check for if the list is empty.

Fixes #6